requirements.txt
# pip install -r requirements.txt
langgraph
langchain
langchain-openai
langchain-ollama
langchain-chroma
langchain-core
pydantic
typing-extensions
numpy
matplotlib
pandapower
numba
python-dotenv
rich
yaspin
sentence-transformers
torch
huggingface-hub
flask
flask-cors